// src/app/component/thi-hanh-an/model/penalty.ts
export interface Penalty {
  penaltyId?: string;
  numYear?: number;
  numMonth?: number;
  numDay?: number;
  disinherit?: string;
  confiscation?: string;
  numMoney?: number | string;
  tichThuTaiSan?: boolean;
  trucXuat?: boolean;
  suspended?: string;
  isCuongChe?: boolean;
  isCheck?: boolean;
}